add "go to sources" toolbar button to profiler

+++ src/cpp/session/modules/SessionProfiler.R
@@ -107,7 +107,10 @@
          profvis <- profilerOptions$profvis
          htmlwidgets::saveWidget(profvis, htmlFile, selfcontained = TRUE)
 
-         file.remove(profilerOptions$profvis$x$message$prof_output)
+         if (resources$tempPath == substr(profilerOptions$profvis$x$message$prof_output, 1, nchar(resources$tempPath)))
+         {
+            file.remove(profilerOptions$profvis$x$message$prof_output)
+         }
       }
 
       return(list(
@@ -173,6 +176,18 @@
 
       return(list(
       ))
+   }, error = function(e) {
+      return(list(error = .rs.scalar(e$message)))
+   })
+})
+
+.rs.addJsonRpcHandler("profile_sources", function(filePath)
+{
+   tryCatch({
+      fixedPath <- filePath
+      exists <- file.exists(fixedPath)
+
+      return(.rs.scalar(if (exists) fixedPath else ""))
    }, error = function(e) {
       return(list(error = .rs.scalar(e$message)))
    })
